src/requirements_pip.txt
accelerate==1.4.0
adbc-driver-manager==0.8.0
adbc-driver-sqlite==0.8.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
altair==5.2.0
annotated-types==0.6.0
anyio==4.8.0
appdirs==1.4.4
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1733250440834/work
attrs==25.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
bitsandbytes==0.45.2
bleach==6.2.0
blinker==1.7.0
blis==1.2.0
cachetools==5.3.2
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.8
cloudpathlib==0.20.0
cloudpickle==3.0.0
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1733502965406/work
confection==0.1.5
connectorx==0.3.2
contourpy==1.2.0
cycler==0.12.1
cymem==2.0.11
dataclasses-json==0.6.7
debugpy @ file:///home/conda/feedstock_root/build_artifacts/debugpy_1737269753336/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1733236420667/work
defusedxml==0.7.1
deltalake==0.13.0
docopt==0.6.2
en_core_web_lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.8.0/en_core_web_lg-3.8.0-py3-none-any.whl#sha256=293e9547a655b25499198ab15a525b05b9407a75f10255e405e8c3854329ab63
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et-xmlfile==1.1.0
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1733208806608/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1733569351617/work
fastjsonschema==2.19.0
filelock==3.17.0
fonttools==4.45.1
frozendict==2.3.10
frozenlist==1.5.0
fsspec==2023.12.2
gevent==23.9.1
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.0.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.28.1
idna==3.10
importlib-metadata==6.11.0
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1719845459717/work
ipython==8.12.3
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1733300866624/work
Jinja2==3.1.5
joblib==1.3.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1733440914442/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1727163409502/work
jupyterlab_pygments==0.3.0
kiwisolver==1.4.5
langchain==0.3.19
langchain-community==0.3.18
langchain-core==0.3.37
langchain-text-splitters==0.3.6
langcodes==3.5.0
langdetect==1.0.9
langsmith==0.3.9
language_data==1.3.0
llvmlite==0.44.0
lxml==4.9.3
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1733416936468/work
mdurl==0.1.2
mistune==3.1.2
mmhash3==3.0.1
more-itertools==10.6.0
mpmath==1.3.0
multidict==6.1.0
murmurhash==1.0.12
mypy-extensions==1.0.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.9.2
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1733325553580/work
networkx==3.4.2
nltk==3.8.1
numba==0.61.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.8.61
nvidia-nvtx-cu12==12.1.105
openai-whisper==20240930
openpyxl==3.0.10
orjson==3.10.15
outcome==1.3.0.post0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1733203243479/work
pandas==2.1.1
pandocfilters==1.5.1
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1733271261340/work
peewee==3.17.0
peft==0.5.0
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1733301927746/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1733327343728/work
Pillow==10.1.0
pipreqs==0.5.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1733232627818/work
plotly==5.18.0
polars==0.19.19
preshed==3.0.9
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1737453357274/work
propcache==0.3.0
protobuf==4.25.1
psutil @ file:///home/conda/feedstock_root/build_artifacts/psutil_1735327328223/work
psycopg2-binary==2.9.10
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1733302279685/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl#sha256=92c32ff62b5fd8cf325bec5ab90d7be3d2a8ca8c8a3813ff487a8d2002630d1f
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1733569405015/work
pyarrow==11.0.0
PyAudio==0.2.11
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydeck==0.8.1b0
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1736243443484/work
pyiceberg==0.5.1
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1733215673016/work
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0
pyzmq @ file:///home/conda/feedstock_root/build_artifacts/pyzmq_1738270958831/work
RapidFuzz==3.12.1
referencing==0.32.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.13.2
safetensors==0.5.2
schedule==1.2.1
scipy==1.11.4
selenium==4.30.0
sentencepiece==0.2.0
shellingham==1.5.4
six @ file:///home/conda/feedstock_root/build_artifacts/six_1733380938961/work
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.23
srsly==2.5.1
srt==3.5.3
stack_data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1733569443808/work
streamlit==1.29.0
strictyaml==1.7.3
sympy==1.13.3
tenacity==8.5.0
thinc==8.3.4
tiktoken==0.9.0
tinycss2==1.4.0
tokenizers==0.13.3
toml==0.10.2
toolz==1.0.0
torch==2.1.1
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1732615904614/work
tqdm==4.67.1
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1733367359838/work
transformers==4.32.0
trio==0.29.0
trio-websocket==0.12.2
triton==2.1.0
typer==0.15.1
typing-inspect==0.9.0
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1733188668063/work
tzdata==2025.1
tzlocal==5.3
urllib3==2.3.0
vaderSentiment==3.3.2
validators==0.34.0
vosk==0.3.45
wasabi==1.1.3
watchdog==6.0.0
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1733231326287/work
weasel==0.4.1
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0
wrapt==1.17.2
wsproto==1.2.0
yarg==0.1.9
yarl==1.18.3
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1732827521216/work
zope.event==5.0
zope.interface==7.2
zstandard==0.23.0
